@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion '25.0.0'
+    compileSdk 32
+    buildToolsVersion '31.0.0'
 
     defaultConfig {
         applicationId "com.example.android.scrolleventsdemo"
-        minSdkVersion 21
-        targetSdkVersion 22
+        minSdk 24
+        targetSdk 32
         versionCode 1
         versionName "1.0"
     }
@@ -18,10 +18,6 @@ android {
         }
     }
 
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
-    }
 }
 
 repositories {
@@ -32,8 +28,8 @@ repositories {
 }
 
 dependencies {
-    compile 'com.android.support:design:22.2.0'
-    compile 'com.android.support:appcompat-v7:22.2.0'
-    compile 'com.android.support:recyclerview-v7:22.2.0'
-    compile fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'com.google.android.material:material:1.6.1'
+    implementation 'androidx.appcompat:appcompat:1.4.2'
+    implementation 'androidx.recyclerview:recyclerview:1.2.1'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
 }
